Better tests for saving/loading

+++ tests/test_continuous.py
@@ -40,6 +40,7 @@ def test_model_manipulation(request, model_class):
 
     :param model_class: (BaseRLModel) A model
     """
+    model_fname = None
     try:
         env = DummyVecEnv([lambda: IdentityEnvBox(eps=0.5)])
 
@@ -47,7 +48,9 @@ def test_model_manipulation(request, model_class):
         model = model_class(policy="MlpPolicy", env=env, seed=0)
         model.learn(total_timesteps=NUM_TIMESTEPS)
 
-        acc_reward, _ = evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
+        env.reset()
+        observations = np.concatenate([env.step([env.action_space.sample()])[0] for _ in range(10)], axis=0)
+        selected_actions, _ = model.predict(observations, deterministic=True)
 
         # saving
         model_fname = './test_model_{}.zip'.format(request.node.name)
@@ -58,11 +61,13 @@ def test_model_manipulation(request, model_class):
         # loading
         model = model_class.load(model_fname)
 
+        # check if model still selects the same actions
+        new_selected_actions, _ = model.predict(observations, deterministic=True)
+        assert np.allclose(selected_actions, new_selected_actions, 1e-4)
+
         # changing environment (note: this can be done at loading)
         env = DummyVecEnv([lambda: IdentityEnvBox(eps=0.5)])
         model.set_env(env)
-
-        loaded_acc_reward, _ = evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
 
         obs = env.reset()
         with pytest.warns(None) as record:
@@ -95,38 +100,17 @@ def test_model_manipulation(request, model_class):
             actions_logprobas = model.action_probability(observations, actions=actions, logp=True)
             assert np.allclose(actions_probas, np.exp(actions_logprobas)), (actions_probas, actions_logprobas)
 
-        # assert <15% diff
-        assert abs(acc_reward - loaded_acc_reward) / max(acc_reward, loaded_acc_reward) < 0.15, \
-            "Error: the prediction seems to have changed between loading and saving"
-
         # learn post loading
         model.learn(total_timesteps=100)
 
-        # validate no reset post learning
-        # This test was failing from time to time for no good reason
-        # other than bad luck
-        # We should change this test
-        # loaded_acc_reward = 0
-        # set_global_seeds(0)
-        # obs = env.reset()
-        # for _ in range(N_EVAL_EPISODES):
-        #     action, _ = model.predict(obs)
-        #     obs, reward, _, _ = env.step(action)
-        #     loaded_acc_reward += reward
-        # loaded_acc_reward = sum(loaded_acc_reward) / N_EVAL_EPISODES
-        # # assert <10% diff
-        # assert abs(acc_reward - loaded_acc_reward) / max(acc_reward, loaded_acc_reward) < 0.1, \
-        #     "Error: the prediction seems to have changed between pre learning and post learning"
-
         # predict new values
-
         evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
 
         # Free memory
         del model, env
 
     finally:
-        if os.path.exists(model_fname):
+        if model_fname is not None and os.path.exists(model_fname):
             os.remove(model_fname)
 
 
@@ -146,7 +130,7 @@ def test_ddpg_eval_env():
     model = DDPG("MlpPolicy", "Pendulum-v0", nb_rollout_steps=5,
                 nb_train_steps=2, nb_eval_steps=10,
                 eval_env=eval_env, verbose=0)
-    model.learn(300)
+    model.learn(NUM_TIMESTEPS)
 
 
 def test_ddpg_normalization():
@@ -157,7 +141,7 @@ def test_ddpg_normalization():
     model = DDPG('MlpPolicy', 'Pendulum-v0', memory_limit=50000, normalize_observations=True,
                  normalize_returns=True, nb_rollout_steps=128, nb_train_steps=1,
                  batch_size=64, param_noise=param_noise)
-    model.learn(300)
+    model.learn(NUM_TIMESTEPS)
     obs_rms_params = model.sess.run(model.obs_rms_params)
     ret_rms_params = model.sess.run(model.ret_rms_params)
     model.save('./test_ddpg.zip')
@@ -185,4 +169,4 @@ def test_ddpg_popart():
     model = DDPG('MlpPolicy', 'Pendulum-v0', memory_limit=50000, normalize_observations=True,
                  normalize_returns=True, nb_rollout_steps=128, nb_train_steps=1,
                  batch_size=64, action_noise=action_noise, enable_popart=True)
-    model.learn(300)
+    model.learn(NUM_TIMESTEPS)
